rename StatisticTable folder
