+++ src/webapp/components/table/statistic-table/MedianRow.tsx
