+++ src/webapp/components/table/statistic-table/PercentTargetMetRow.tsx
