+++ src/webapp/components/table/statistic-table/StatisticTable.tsx
